# packages/api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/maritime/wharfs/{id}": {
      "get": {
        "tags": [
          "涉海档案 · 码头"
        ],
        "summary": "查询码头完整档案",
        "operationId": "detail_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWharfView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "涉海档案 · 码头"
        ],
        "summary": "编辑或启停码头，需最新版本",
        "operationId": "update_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteWharf"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWharfView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "涉海档案 · 码头"
        ],
        "summary": "删除未被引用的码头，需最新版本",
        "operationId": "delete_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/maritime/ports/{id}": {
      "get": {
        "tags": [
          "涉海档案 · 港口"
        ],
        "summary": "查询港口完整档案",
        "operationId": "detail_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePortView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "涉海档案 · 港口"
        ],
        "summary": "编辑或启停港口，需最新版本",
        "operationId": "update_6",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WritePort"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePortView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "涉海档案 · 港口"
        ],
        "summary": "删除未被引用的港口，需最新版本",
        "operationId": "delete_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/maritime/wharfs": {
      "get": {
        "tags": [
          "涉海档案 · 码头"
        ],
        "summary": "分页查询码头",
        "operationId": "list_5",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "portId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "policeStationId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "responsibleOfficerId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResultWharfView"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "涉海档案 · 码头"
        ],
        "summary": "新建码头",
        "operationId": "create_6",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteWharf"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWharfView"
                }
              }
            }
          }
        }
      }
    },
    "/api/maritime/ports": {
      "get": {
        "tags": [
          "涉海档案 · 港口"
        ],
        "summary": "分页查询港口",
        "operationId": "list_6",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResultPortView"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "涉海档案 · 港口"
        ],
        "summary": "新建港口",
        "operationId": "create_7",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WritePort"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePortView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
...
      "WriteWharf": {
        "type": "object",
        "properties": {
          "version": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "新建不传；编辑必填，缺失返回400，旧版本返回409",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "region": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "location": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200,
            "minLength": 0
          },
          "status": {
            "type": "string",
            "minLength": 1,
            "pattern": "ENABLED|DISABLED"
          },
          "purpose": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200,
            "minLength": 0
          },
          "portId": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "[1-9][0-9]{0,18}"
          },
          "policeStationId": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "[1-9][0-9]{0,18}"
          },
          "responsibleOfficerId": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "[1-9][0-9]{0,18}"
          }
        },
        "required": [
          "name",
          "region",
          "status"
        ]
      },
      "ApiResponseWharfView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/WharfView"
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      },
      "WharfView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "location": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "purpose": {
            "type": [
              "string",
              "null"
            ]
          },
          "portId": {
            "type": [
              "string",
              "null"
            ]
          },
          "policeStationId": {
            "type": [
              "string",
              "null"
            ]
          },
          "responsibleOfficerId": {
            "type": [
              "string",
              "null"
            ]
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "fixtureKey": {
            "type": [
              "string",
              "null"
            ]
          },
          "portName": {
            "type": [
              "string",
              "null"
            ]
          },
          "portStatus": {
            "type": [
              "string",
              "null"
            ]
          },
          "policeStationName": {
            "type": [
              "string",
              "null"
            ]
          },
          "policeStationStatus": {
            "type": [
              "string",
              "null"
            ]
          },
          "responsibleOfficerName": {
            "type": [
              "string",
              "null"
            ]
          },
          "responsibleOfficerStatus": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "createdAt",
          "fixtureKey",
          "id",
          "location",
          "name",
          "policeStationId",
          "policeStationName",
          "policeStationStatus",
          "portId",
          "portName",
          "portStatus",
          "purpose",
          "region",
          "responsibleOfficerId",
          "responsibleOfficerName",
          "responsibleOfficerStatus",
          "status",
          "updatedAt",
          "version"
        ]
      },
      "WritePort": {
        "type": "object",
        "properties": {
          "version": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "新建不传；编辑必填，缺失返回400，旧版本返回409",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "region": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "location": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200,
            "minLength": 0
          },
          "status": {
            "type": "string",
            "minLength": 1,
            "pattern": "ENABLED|DISABLED"
          },
          "purpose": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200,
            "minLength": 0
          }
        },
        "required": [
          "name",
          "region",
          "status"
        ]
      },
      "ApiResponsePortView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PortView"
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      },
      "PortView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "location": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "purpose": {
            "type": [
              "string",
              "null"
            ]
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "fixtureKey": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "createdAt",
          "fixtureKey",
          "id",
          "location",
          "name",
          "purpose",
          "region",
          "status",
          "updatedAt",
          "version"
        ]
      },
...
      "ApiResponsePageResultWharfView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResultWharfView"
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      },
      "PageResultWharfView": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WharfView"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
      "ApiResponsePageResultPortView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResultPortView"
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      },
      "PageResultPortView": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortView"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
...
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      }
    }
  }
}